Fix a few bugs

+++ Models/StockPurchaseRequest.cs
@@ -3,6 +3,6 @@ namespace StocksCoreApi.Models
     public class StockPurchaseRequest
     {
         public string StockName { get; set; }
-        public int Units { get; set; }
+        public decimal Units { get; set; }
     }
 }
